+++ src/hierarchy-tree-selector.component.jsx
@@ -83,6 +83,8 @@ export default class HierarchyTreeSelector extends React.PureComponent {
     valueKey: PropTypes.string,
     childKey: PropTypes.string,
     lockedKey: PropTypes.string,
+    leafValueKey: PropTypes.string,
+    sortKey: PropTypes.string,
     treeData: PropTypes.arrayOf(PropTypes.shape({})),
     grid: gridShape.isRequired,
     gridColumns: PropTypes.arrayOf(gridColumnShape).isRequired,
@@ -107,6 +109,8 @@ export default class HierarchyTreeSelector extends React.PureComponent {
     valueKey: 'name',
     childKey: 'children',
     lockedKey: undefined,
+    leafValueKey: undefined,
+    sortKey: undefined,
     treeData: [],
     className: '',
     translations: defaultTranslations,
@@ -128,8 +132,8 @@ export default class HierarchyTreeSelector extends React.PureComponent {
     };
   }
 
-  componentDidUpdate(prevProps) {
-    if (this.props.defaultExpandedKeys.length > 0 && prevProps.defaultExpandedKeys.length === 0) {
+  componentDidMount() {
+    if (this.props.defaultExpandedKeys.length > 0) {
       this.onExpand(this.props.defaultExpandedKeys);
     }
   }
@@ -404,9 +408,12 @@ export default class HierarchyTreeSelector extends React.PureComponent {
    */
   setDataToGrid = (items, setNewItems = false) => {
     let data = List();
-    const { grid, gridColumns, gridData } = this.props;
+    const {
+      grid, gridColumns, gridData, sortKey,
+    } = this.props;
     if (!setNewItems) data = gridData.slice();
-    const newGridItems = data.concat(items);
+    let newGridItems = data.concat(items);
+    if (sortKey) newGridItems = newGridItems.sortBy(i => i.get(sortKey));
 
     this.props.setData(grid, gridColumns, newGridItems);
     this.props.clearSelectedItems(grid);
@@ -503,6 +510,7 @@ export default class HierarchyTreeSelector extends React.PureComponent {
   render() {
     const {
       valueKey,
+      leafValueKey,
       idKey,
       treeData,
       grid,
@@ -523,6 +531,7 @@ export default class HierarchyTreeSelector extends React.PureComponent {
               treeData={treeData}
               dataLookUpKey={idKey}
               dataLookUpValue={valueKey}
+              dataLookUpLeafValue={leafValueKey}
               dataLookUpChildren={childKey}
               onSelect={this.onTreeItemSelect}
               onExpand={this.onExpand}